requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
icalendar==4.0.1
idna==2.6
pkg-resources==0.0.0
pyCalverter==1.6.1
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
